הוספתי שינויים מסוימים

+++ The-piano-house.Api/Controllers/CustomersController.cs
@@ -1,10 +1,12 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Mvc;
+using The_piano_house.Api.Models;
 using The_piano_house.Core.Entities;
 using The_Piano_house.Core.Services;
 using The_Piano_house.Servise;
+using The_Piano_house.Core.DTOs;
 
 
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace The_piano_house.Api.Controllers
 {
@@ -13,42 +15,48 @@ namespace The_piano_house.Api.Controllers
     public class CustomersController : ControllerBase
     {
         private readonly ICustomersService _customersService;
-        public CustomersController(ICustomersService customersService)
+        private readonly  IMapper _mapper;
+        public CustomersController(ICustomersService customersService, IMapper mapper)
         {
             _customersService = customersService;
+            _mapper = mapper;
         }
 
 
         // GET: api/<CustomersController>
         [HttpGet]
-        public List<Customers> Get()
+        public ActionResult Get()
         {
-            return  _customersService.Get();
+            return Ok(_customersService.Get());
         }
 
 
         // GET api/<CustomersController>/5
         [HttpGet("{id}")]
-       // public ActionResult< Customers> Get(int id)
-        public  Customers  Get(int id)
+        // public ActionResult< Customers> Get(int id)
+        public ActionResult Get(int id)
         {
-            return _customersService.Get(id);
-           // var ev = _conect.CustomersList.Find(e => e.id == id);
+            return Ok(_customersService.Get(id));
+            // var ev = _conect.CustomersList.Find(e => e.id == id);
 
-         //   if (ev == null)
-           //     return NotFound();
+            //   if (ev == null)
+            //     return NotFound();
 
-           //     return ev;
+            //     return ev;
         }
 
 
         // POST api/<CustomersController>
         [HttpPost]
-      //  public ActionResult Post([FromBody] Customers c)
-        public void Post([FromBody] Customers c)
+        //  public ActionResult Post([FromBody] Customers c)
+        public  ActionResult Post([FromBody] CustomerPostModel c)
         {
-          _customersService.Post(c);  
-           // return Ok();
+            var CustomerToAdd = new Customer { Name = c.Name, Phone=c.Phone,Address=c.Address };
+
+            var newCust=_customersService.Post(CustomerToAdd);
+            var custumerDto = _mapper.Map<CustumerDTO>(newCust);
+            return Ok(CustomerToAdd);
+            // return Ok();
             //if (c.id.ToString().Length != 9)
             //    return BadRequest();
             //_conect.CustomersList.Add(new Customers { id =c.id,name=c.name,phone=c.phone,address=c.address,lastPurchaseDate=c.lastPurchaseDate  });
@@ -58,32 +66,42 @@ namespace The_piano_house.Api.Controllers
 
         // PUT api/<CustomersController>/5
         [HttpPut("{id}")]
-       // public ActionResult Put(int id, [FromBody] Customers c)
-        public void Put(int id, [FromBody] Customers c)
+        // public ActionResult Put(int id, [FromBody] Customers c)
+        public ActionResult Put(int id, [FromBody] CustomerPostModel c)
         {
+            Customer c1 = new Customer();
+            c1.Name = c.Name;
+            c1.Phone = c.Phone;
+            c1.Address = c.Address;
+            var newc=_customersService.Put(id, c1);
+            var custumerDto = _mapper.Map<CustumerDTO>(newc);
+            return Ok(c1);
 
-
-           _customersService.Put(id, c);    
-           // if (ev == null)
+            // if (ev == null)
             //    return NotFound();
-           
-         
-            
-           //  return Ok();
+
+
+
+            //  return Ok();
         }
 
         // DELETE api/<CustomersController>/5
         [HttpDelete("{id}")]
-     //   public ActionResult Delete(int id)
-        public void Delete(int id)
+        //   public ActionResult Delete(int id)
+        public ActionResult Delete(int id)
         {
-            
+            var cust = _customersService. Get(id);
+            if (cust is null)
+            {
+                return NotFound();
+            }
             _customersService.Delete(id);
+            return NoContent();
             //if (ev == null)
-             //   return NotFound();             
-                //  return Ok();      
-                
-            
+            //   return NotFound();             
+            //  return Ok();      
+
+
         }
     }
 }
