+++ The-piano-house.Api/Controllers/ProviderController.cs
@@ -1,7 +1,11 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Mvc;
 using The_piano_house.Core.Entities;
+using The_Piano_house.Core.DTOs;
 using The_Piano_house.Core.Services;
 using The_Piano_house.Servise;
+using The_piano_house.Api.Models;
+
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -12,65 +16,78 @@ namespace The_piano_house.Api.Controllers
     public class ProviderController : ControllerBase
     {
         private readonly IProviderService _providerService;
-        public ProviderController(IProviderService providerService)
+        private readonly IMapper _mapper;
+
+        public ProviderController(IProviderService providerService , IMapper mapper)
         {
             _providerService = providerService;
+            _mapper = mapper;
         }
 
         // GET: api/<ProviderController>
         [HttpGet]
-        public List<Provider> Get()
+        public ActionResult<Provider> Get()
         {
-            return _providerService.Get();
+            return Ok( _providerService.Get());
         }
 
         // GET api/<ProviderController>/5
         [HttpGet("{id}")]
-        public ActionResult< Provider> Get(int id)
+        public ActionResult<Provider> Get(int id)
         {
-          
-        return    _providerService.Get(id);
-            //if (ev == null)
-             //   return NotFound();
 
-           //return ev;
+            return Ok( _providerService.Get(id));
+            //if (ev == null)
+            //   return NotFound();
+
+            //return ev;
         }
         // POST api/<ProviderController>
         [HttpPost]
-        public ActionResult Post([FromBody] Provider p)
+        public ActionResult Post([FromBody] ProviderPostModel p)
         {
-           // if (p.id.ToString().Length != 9)
+            var pToadd = new Provider { Name = p.Name, Phone = p.Phone, Address = p.Address };
+
+            var newp = _providerService.Post(pToadd); 
+            var custumerDto = _mapper.Map<ProviderDTO>(newp);
+            return Ok(pToadd);
+            // if (p.id.ToString().Length != 9)
             //    return BadRequest();
-            _providerService.Post(p);   
-            return Ok();
+           
+           
         }
 
         // PUT api/<ProviderController>/5
         [HttpPut("{id}")]
-        public ActionResult Put(int id, [FromBody] Provider p)
+        public ActionResult Put(int id, [FromBody] ProviderPostModel p)
         {
 
-           
+            Provider p1 = new Provider();
+            p1.Name = p.Name;
+            p1.Phone = p.Phone;
+            p1.Address = p.Address;
+            var newp = _providerService.Put(id, p1);
+            var custumerDto = _mapper.Map<ProviderDTO>(newp);
 
             //if (ev == null)
             //   return NotFound();
-               _providerService.Put(id, p);
-            return Ok();
-                
-            
+            return Ok(p1 );
         }
 
         // DELETE api/<ProviderController>/5
         [HttpDelete("{id}")]
         public ActionResult Delete(int id)
         {
-            
+            var pro = _providerService.Get(id);
+            if (pro is null)
+            {
+                return NotFound();
+            }
 
-           // if (ev == null)
+            _providerService.Delete(id);
+            return NoContent();
+            // if (ev == null)
             //    return NotFound();
-           _providerService.Delete(id);
-            return Ok();
-
         }
     }
 }
